requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
betterproto==1.2.5
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
discord.py==2.2.3
Django==4.2
djangorestframework==3.14.0
frozenlist==1.3.3
grpclib==0.4.3
h2==4.1.0
hpack==4.0.0
http-ece==1.1.0
hyperframe==6.0.1
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.3
oscrypto==1.3.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
protobuf==4.22.3
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
rustplus==5.5.13
rustPlusPushReceiver==0.4.1
scipy==1.10.1
sqlparse==0.4.4
stringcase==1.2.0
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
websocket-client==1.5.1
whitenoise==6.4.0
yarl==1.9.2
